docker: Setup docker environment
Docker container has been configured with the Debian 11 image following
the convention currently in place at LNLS, and its subversion has been
chosen to be the latest one.

EPICS 7.0.7 has been chosen to be the base version so that the build can
use the latest modules from EPICS. However, `pvAccess` is not enabled,
meaning that the used communication protocol is still Channel Access.
This will allow us to use the `int64in` and `int64out` record types from
EPICS 3.16+.

Only the needed modules from area detector have been cloned and
compiled, so that the resulting image contains only the necessary
dependencies for running.

Similarly, ADSupport includes several libraries which submodules might
depend on. By default, all of them are built. However, this is not
needed for all IOCs. Therefore, here we configure it to only build the
required libraries, which is the `xml2` lib required by ADCore.