database: Change driver and plugins prefix to use colon
It is a pattern in areaDetector to use the plugins and driver names
suffixed by a colon, for instance `Cam1:` instead of simply `Cam1`.
In configuration files from the beamlines, this pattern is already used
by rewriting the `st.cmd` file provided from the original repository.
Such names currently only impact the aliases and forward links made in
the database following Sirius PV naming standard, which is used only in
the accelerator.

Therefore, a downside of following this approach is having other modules
that use directly the PVs from area detector and its plugins
reconfigured. An alternative to solve this is using a macro that will be
expanded based on the configuration provided. However, a better study on
whether the PV from the standard are really needed should be done
beforehand.